Separated runOnSchemaChange from AbstractPowerSyncData.
